RavenDB-4132: Root Objects headers should be more uniform.
Now there is an explicit shared RootObject header we can use instead of casting to a TreeRootObject to know the type of the object. Also moved the shared object from an internal implementation detail to the Voron.Data namespace.

+++ src/Voron/Impl/Scratch/ScratchBufferPool.cs
@@ -82,7 +82,7 @@ namespace Voron.Impl.Scratch
         {
             if (tx == null)
                 throw new ArgumentNullException(nameof(tx));
-            var size = (int)Bits.NextPowerOf2(numberOfPages);
+            var size = Bits.NextPowerOf2(numberOfPages);
 
             PageFromScratchBuffer result;
             var current = _current;
